add cloud api support
