+++ cloud_examples/local_example.py
