+++ pyfronius/errors.py
@@ -1,0 +1,6 @@
+class NotAuthorizedException(ValueError):
+    pass
+
+
+class NotFoundException(ValueError):
+    pass
