+++ pyfronius/schema/error.py
@@ -1,0 +1,8 @@
+from datetime import datetime
+
+from pydantic import BaseModel
+
+
+class Error(BaseModel):
+    responseError: str
+    responseMessage: str
