--- pyfronius/tests/test_structure/v1/solar_api/v1/GetMeterRealtimeData.cgi?Scope=System
@@ -1,31 +1,0 @@
-{
-	"Head" : {
-		"RequestArguments" : {
-			"DeviceClass" : "Meter",
-			"Scope" : "System"
-		},
-		"Status" : {
-			"Code" : 0,
-			"Reason" : "",
-			"UserMessage" : ""
-		},
-		"Timestamp" : "2019-01-10T23:33:13+01:00"
-	},
-	"Body" : {
-		"Data" : {
-			"0" : {
-				"Details" : {
-					"Serial" : "",
-					"Model" : "",
-					"Manufacturer" : "Fronius"
-				},
-				"TimeStamp" : 1547159592,
-				"Enable" : 1,
-				"Visible" : 1,
-				"PowerReal_P_Sum" : -367.722145,
-				"Meter_Location_Current" : 1,
-				"EnergyReal_WAC_Minus_Relative" : 17
-			}
-		}
-	}
-}
