--- pyfronius/tests/test_structure/v1/solar_api/v1/GetStorageRealtimeData.cgi?Scope=Device&DeviceId=0
@@ -1,35 +1,0 @@
-{
-   "Body" : {
-      "Data" : {
-         "Controller" : {
-            "Capacity_Maximum" : 11520,
-            "DesignedCapacity" : 11520,
-            "Details" : {
-               "Manufacturer" : "BYD",
-               "Model" : "BYD Battery-Box HV",
-               "Serial" : "123456789-12345"
-            },
-            "Enable" : 1,
-            "StateOfCharge_Relative" : 8,
-            "Status_BatteryCell" : 0,
-            "Temperature_Cell" : 19.699999999999999,
-            "TimeStamp" : 1641402599,
-            "Voltage_DC" : 0
-         },
-         "Modules" : []
-      }
-   },
-   "Head" : {
-      "RequestArguments" : {
-         "DeviceClass" : "Storage",
-         "DeviceId" : "0",
-         "Scope" : "Device"
-      },
-      "Status" : {
-         "Code" : 0,
-         "Reason" : "",
-         "UserMessage" : ""
-      },
-      "Timestamp" : "2022-01-05T18:10:02+01:00"
-   }
-}
